use temporary directory for jobs
Separates the configuration directory from the queue directory, to be able to
use a temporary directory for the queue files.

Using a temporary directory for queue files is a good idea because some devices
may be using a read-only file system, e.g. when using a Raspberry Pi with an
SD card. In that case it will not be possible to write to locations in the home
directory or similar.

Writing to the configuration directory is not necessary during operation, but
writing to the queue directory is definitely necessary, so it makes sense to
separate the two.