validate jobs before running
This validates the job with the currently activated passes, just before
starting the actual job. This also takes into account the current offset
of the machine.

Because of running this late, this allows to take into account some
more specific use cases like for example using the machine offset to
set the center of a job rather than setting the top left corner of a job
with strategically putting things in negative absolute coordinates.

Adding a default to show the error message in the user interface in
frontend/Request.js is necessary because things like jogging/moving in
the frontend also send "jobs" that might fail due to this verification.